fixed the float problem

+++ csharp-prep/Prep4/Program.cs
@@ -13,7 +13,7 @@ class Program
         int response=1;
         int total=0;
         int highest=0;
-        int average=0;
+        float average=0;
         while (response!=0)
         {
             Console.WriteLine("Type a number to add to the list.");
@@ -33,7 +33,7 @@ class Program
             }
             else
             {
-                average=total/numbers.Count;
+                average=(float)total/numbers.Count;
                 Console.WriteLine ($"sum total is {total}");
                 Console.WriteLine ($"Highest number is {highest}");
                 Console.WriteLine ($"Average of numbers is {average}");
